ticket:9101: Fixed a bug in CollationBuilder, which resolved a failing RuleBasedCollator construction in TestDataDriven.

+++ collv2/main/classes/collate/src/com/ibm/icu/impl/coll/CollationRootElements.java
@@ -383,7 +383,7 @@ public final class CollationRootElements {
             ++index;
             terLimit = getTertiaryBoundary();
         }
-        long st = (s << 16) | t;
+        long st = ((long)s << 16) | t;
         for(;;) {
             long secTer = elements[index];
             // No tertiary greater than t for this primary+secondary.
